Added Version 1.0 Version 1.0: The barebones HTML and CSS of the page. This version includes the various html, css, and image resources needed to complete the google.com recreation project. So far, the page includes a navbar element; the google logo, search bar and 2 buttons; and a footer bar.
The next version will see the footer bar postioned properly at the bottom of the page, and styling tweaks to the various elements.

Self Reflection:
The project started off smoothly but I started slowing down once I began positioning the elements. In the future, I should take more time to design a layout BEFORE starting
to code the page. By doing this, I would have a guide to follow instead of trying to design a layout on the fly.